chore(util): migrate util to functions, it is not a service
