+++ packages/server/src/commands/createFolders.ts
@@ -1,6 +1,6 @@
 import filetree from "../services/filetree/index.js";
-import log from "../services/log.js";
-import utils from "../services/utils.js";
+import log from "../services/log/index.js";
+import { utils } from "../utils/index.js";
 import type { CommandHandler } from "./index.js";
 
 interface CreateFoldersMessage {
