+++ packages/server/src/commands/deleteFile.ts
@@ -1,5 +1,5 @@
 import filetree from "../services/filetree/index.js";
-import log from "../services/log.js";
+import log from "../services/log/index.js";
 
 import type { CommandHandler } from "./index.js";
