+++ packages/server/src/commands/getMedia.ts
@@ -2,8 +2,8 @@ import path from "node:path";
 import { imageSize } from "image-size";
 
 import filetree from "../services/filetree/index.js";
-import log from "../services/log.js";
-import utils from "../services/utils.js";
+import log from "../services/log/index.js";
+import { utils } from "../utils/index.js";
 
 import type { CommandHandler } from "./index.js";
