+++ packages/server/src/commands/saveFile.ts
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import filetree from "../services/filetree/index.js";
-import log from "../services/log.js";
+import log from "../services/log/index.js";
 
 import type { CommandHandler } from "./index.js";
