+++ packages/server/src/index.ts
@@ -1,16 +1,17 @@
-import { droppy } from "./services/server.js";
+import { droppy } from "./services/server/index.js";
 
 export { default as cfg } from "./services/cfg/index.js";
 export { default as cookies } from "./services/cookies/index.js";
 export { default as csrf } from "./services/csrf/index.js";
-export { default as db } from "./services/db/db.js";
+export { default as db } from "./services/db/index.js";
 export { default as filetree } from "./services/filetree/index.js";
-export { default as log } from "./services/log.js";
-export { default as manifest } from "./services/manifest.js";
-export { default as paths } from "./services/paths.js";
-export { default as resources } from "./services/resources.js";
-export { default as svg } from "./services/svg.js";
+export { default as log } from "./services/log/index.js";
+export { default as manifest } from "./services/manifest/index.js";
+export { default as paths } from "./services/paths/index.js";
+export { default as resources } from "./services/resources/index.js";
+export { default as svg } from "./services/svg/index.js";
 export { default as users } from "./services/users/index.js";
-export { default as utils } from "./services/utils.js";
+
+export { utils } from "./utils/index.js";
 
 export { droppy };
