+++ packages/server/src/utils/filesystem/getDispo.ts
@@ -1,0 +1,8 @@
+import path from "node:path";
+import cd from "content-disposition";
+
+export default function getDispo(fileName: string, download: boolean) {
+    return cd(path.basename(fileName), {
+        type: download ? "attachment" : "inline",
+    });
+}
