+++ packages/server/src/utils/filesystem/normalizePath.ts
@@ -1,0 +1,3 @@
+export default function normalizePath(path: string) {
+    return path.replace(/[\\|/]+/g, "/");
+}
