+++ packages/server/src/utils/filesystem/readFile.ts
@@ -1,0 +1,20 @@
+import fs from "node:fs";
+
+export default function readFile(p, cb) {
+    if (typeof p !== "string") {
+        return cb(null);
+    }
+
+    fs.stat(p, (_, stats) => {
+        if (stats?.isFile()) {
+            fs.readFile(p, (err, data) => {
+                if (err) {
+                    return cb(err);
+                }
+                cb(null, String(data));
+            });
+        } else {
+            cb(null);
+        }
+    });
+}
