+++ packages/server/src/utils/filesystem/sanitizePathsInString.ts
@@ -1,0 +1,9 @@
+import escapeStringRegexp from "escape-string-regexp";
+import paths from "../../services/paths/index.js";
+
+export default function sanitizePathsInString(str: string) {
+    return (str || "").replace(
+        new RegExp(escapeStringRegexp(paths.get().files), "g"),
+        "",
+    );
+}
