+++ packages/server/src/utils/strings/addUploadTempExt.ts
@@ -1,0 +1,3 @@
+export default function addUploadTempExt(p: string) {
+    return p.replace(/(\/?[^/]+)/, (_, p1) => `${p1}.droppy-upload`);
+}
