+++ packages/server/src/utils/strings/arrify.ts
@@ -1,0 +1,3 @@
+export default function arrify(val) {
+    return Array.isArray(val) ? val : [val];
+}
