+++ packages/server/src/utils/strings/createSid.ts
@@ -1,0 +1,5 @@
+import crypto from "node:crypto";
+
+export default function createSid() {
+    return crypto.randomBytes(64).toString("base64").substring(0, 48);
+}
