+++ packages/server/src/utils/strings/extensionRe.ts
@@ -1,0 +1,8 @@
+import escapeStringRegexp from "escape-string-regexp";
+
+export default function extensionRe(arr: string[]) {
+    const result = arr.map((ext) => {
+        return escapeStringRegexp(ext);
+    });
+    return new RegExp(`\\.(${result.join("|")})$`, "i");
+}
