+++ packages/server/src/utils/strings/pretty.ts
@@ -1,0 +1,9 @@
+import util from "node:util";
+
+export default function pretty(data) {
+    return util
+        .inspect(data, { colors: true })
+        .replace(/^\s+/gm, " ")
+        .replace(/\s+$/gm, "")
+        .replace(/[\r\n]+/gm, "");
+}
