+++ packages/server/src/utils/strings/removeUploadTempExt.ts
@@ -1,0 +1,3 @@
+export default function removeUploadTempExt(p: string) {
+    return p.replace(/(^\/?[^/]+)(\.droppy-upload)/, (_, p1) => p1);
+}
