+++ packages/server/src/utils/strings/rootname.ts
@@ -1,0 +1,3 @@
+export default function rootname(p: string) {
+    return p.split("/").find((p) => !!p);
+}
